Implemented Aquatic Map Modifiers and Populators
Altered all sub classes to use Aquatic Map and altered ArrayGrid3D to be
compatible with my uses

+++ src/main/java/com/csm/rover/simulator/objects/io/jsonserial/RecursiveGridListDeserializer.java
@@ -20,7 +20,7 @@ public class RecursiveGridListDeserializer extends JsonDeserializer<RecursiveGri
         for (int i = 0; i < valuesNode.size(); i++){
             JsonNode valueNode = valuesNode.get(i);
             JsonNode coordsNode = valueNode.get("coordinates");
-            int[] coords = new int[layers];
+            double[] coords = new double[layers];
             for (int j = 0; j < layers; j++){
                 coords[j] = coordsNode.get(j).asInt();
             }
